Adding multi-platform support for linux/amd64 and linux/arm64 (#83)
* Adding multi-platform support for linux/amd64 and linux/arm64

* Bumped GHA versions
Added `--platform` option to `docker build` commands in e2e tests
Added `PLATFORM` variable to Makefile for local builds
Golang version update
Dependency updates


* Need to use `docker buildx` to get platform flags in Dockerfile


* `docker buildx` is apparently not supported in GHA


* Trying to get `BUILDPLATFORM` down to Dockerfiles.


* Defaulted `BUILDPLATFORM`


* Updated tests referring to old v1beta policy
Problem only shows when updating the kubebuilder-tools
  to 1.25.0 (when the v1beta policy was deprecated)


* Updating e2e-tests
Added podman support for e2d tests


* Testing e2e-tests w/`docker buildx build`
Tied to use the `--load` option instead of `--push` but didn't work
May be a Rancher Desktop issue
> error: docker exporter does not currently support exporting manifest lists


* Added logic to handle podman buildx behavior differences
  `podman buildx` auto-loads the image into the local docker repo
  `podman buildx` doesn't support `--push` arg
  `docker pull` is not necessary with podman
Refactored create_dev_cluster.sh


* Updated e2e-test node versions
Added both platforms to docker build commands in e2e-tests


* Can't specify multiple platforms w/regular docker build


* Updated quick-start.md


* Major refactoring of create_dev_cluster.sh
Mainly for testing different platform and build configurations


* Typos and logic


* Typos and logic


* Typos and logic

+++ hack/delete_dev_cluster.sh
@@ -1,5 +1,27 @@
 #!/bin/bash
+
 cluster_name="e2e-tests"
 
+function usage {
+  cat << !
+USAGE: $0 [-c cluster]
+
+Deletes a dev cluster
+
+-c|--cluster   | cluster
+!
+}
+
+while (( "$#" )); do
+  opt="$1"; shift;
+  case "$opt" in
+    "-c"|"--cluster") cluster_name="$1" shift;;
+    *) echo "invalid option: \""$opt"\"" >&2; usage; exit 1;;
+  esac
+done
+
 kind delete cluster --name $cluster_name > /dev/null 2>&1
-rm -f ./e2e-tests-kubeconfig
+
+if [[ -f ./e2e-tests-kubeconfig ]]; then
+ rm -f ./e2e-tests-kubeconfig
+fi
